chore: Move Kuksa Val V1 stuff in v1 folder
