+++ kuksa-java-sdk/src/main/kotlin/org/eclipse/kuksa/connectivity/databroker/v1/request/DataBrokerRequest.kt
@@ -16,17 +16,23 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.eclipse.kuksa.connectivity.databroker.request
+package org.eclipse.kuksa.connectivity.databroker.v1.request
 
-import org.eclipse.kuksa.vsscore.model.VssNode
+import org.eclipse.kuksa.proto.v1.Types.Field
+import org.eclipse.kuksa.proto.v1.Types.Field.FIELD_VALUE
 
 /**
- * Uses [VssNode] models to request information for the [org.eclipse.kuksa.connectivity.databroker.DataBrokerConnection]
- * which are generated by the [org.eclipse.kuksa.vsscore.annotation.VssModelGenerator].
+ * Consists of request information for the [org.eclipse.kuksa.connectivity.databroker.DataBrokerConnection].
  */
-interface VssNodeDataBrokerRequest<T : VssNode> : DataBrokerRequest {
+interface DataBrokerRequest {
     /**
-     * A generated [VssNode] model.
+     * The VehicleSignalSpecification (VSS) path.
      */
-    val vssNode: T
+    val vssPath: String
+
+    /**
+     * The corresponding field type(s) of the [vssPath] request. The [fields] can be used to subscribe to different
+     * information. The default is [FIELD_VALUE].
+     */
+    val fields: Array<out Field>
 }
