+++ kuksa-java-sdk/src/main/kotlin/org/eclipse/kuksa/connectivity/databroker/v1/request/SubscribeRequest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 - 2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,17 @@
  * limitations under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
+ *
  */
 
-package org.eclipse.kuksa.connectivity.databroker.listener
+package org.eclipse.kuksa.connectivity.databroker.v1.request
 
-import org.eclipse.kuksa.pattern.listener.Listener
+import org.eclipse.kuksa.proto.v1.Types
 
 /**
- * The [DisconnectListener] can be registered to
- * [org.eclipse.kuksa.connectivity.databroker.DataBrokerConnection.disconnectListeners]
- * When registered it will notify about manual or unexpected connection disconnects from the DataBroker.
+ * Used for subscribe requests with [org.eclipse.kuksa.connectivity.databroker.v1.DataBrokerConnection.subscribe].
  */
-fun interface DisconnectListener : Listener {
-    /**
-     * Will be triggered, when the connection to the DataBroker was closed manually or unexpectedly.
-     */
-    fun onDisconnect()
-}
+open class SubscribeRequest @JvmOverloads constructor(
+    override val vssPath: String,
+    override vararg val fields: Types.Field = arrayOf(Types.Field.FIELD_VALUE),
+) : DataBrokerRequest
