+++ kuksa-java-sdk/src/main/kotlin/org/eclipse/kuksa/connectivity/databroker/v1/request/VssNodeDataBrokerRequest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 - 2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,20 @@
  * limitations under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
+ *
  */
 
-package org.eclipse.kuksa.connectivity.databroker.request
+package org.eclipse.kuksa.connectivity.databroker.v1.request
 
-import org.eclipse.kuksa.proto.v1.Types
-import org.eclipse.kuksa.proto.v1.Types.Datapoint
+import org.eclipse.kuksa.vsscore.model.VssNode
 
 /**
- * Used for update requests with [org.eclipse.kuksa.connectivity.databroker.DataBrokerConnection.update].
+ * Uses [VssNode] models to request information for the [org.eclipse.kuksa.connectivity.databroker.v1.DataBrokerConnection]
+ * which are generated by the [org.eclipse.kuksa.vsscore.annotation.VssModelGenerator].
  */
-open class UpdateRequest @JvmOverloads constructor(
-    override val vssPath: String,
-    val dataPoint: Datapoint,
-    override vararg val fields: Types.Field = arrayOf(Types.Field.FIELD_VALUE),
-) : DataBrokerRequest
+interface VssNodeDataBrokerRequest<T : VssNode> : DataBrokerRequest {
+    /**
+     * A generated [VssNode] model.
+     */
+    val vssNode: T
+}
