+++ kuksa-java-sdk/src/main/kotlin/org/eclipse/kuksa/connectivity/databroker/v1/request/VssNodeUpdateRequest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 - 2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,22 @@
  * limitations under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
+ *
  */
 
-package org.eclipse.kuksa.connectivity.databroker.request
+package org.eclipse.kuksa.connectivity.databroker.v1.request
 
 import org.eclipse.kuksa.proto.v1.Types
+import org.eclipse.kuksa.vsscore.model.VssNode
 
 /**
- * Used for fetch requests with [org.eclipse.kuksa.connectivity.databroker.DataBrokerConnection.fetch].
+ * Used for update requests with a generated [VssNode] model and
+ * [org.eclipse.kuksa.connectivity.databroker.v1.DataBrokerConnection.update].
  */
-open class FetchRequest @JvmOverloads constructor(
-    override val vssPath: String,
+class VssNodeUpdateRequest<T : VssNode> @JvmOverloads constructor(
+    override val vssNode: T,
     override vararg val fields: Types.Field = arrayOf(Types.Field.FIELD_VALUE),
-) : DataBrokerRequest
+) : VssNodeDataBrokerRequest<T> {
+    override val vssPath: String
+        get() = vssNode.vssPath
+}
