+++ kuksa-java-sdk/src/test/kotlin/org/eclipse/kuksa/connectivity/databroker/DataBrokerConnectorProvider.kt
@@ -27,6 +27,7 @@ import org.eclipse.kuksa.connectivity.authentication.JsonWebToken
 import org.eclipse.kuksa.connectivity.authentication.JwtType
 import org.eclipse.kuksa.connectivity.databroker.docker.DEFAULT_PORT_INSECURE
 import org.eclipse.kuksa.connectivity.databroker.docker.DEFAULT_PORT_SECURE
+import org.eclipse.kuksa.connectivity.databroker.v1.DataBrokerConnector
 import org.eclipse.kuksa.model.TimeoutConfig
 import org.eclipse.kuksa.test.TestResourceFile
 import java.io.IOException
